Split mean solar distance code to it's own file
