+++ appOPHD/MeanSolarDistance.h
@@ -1,0 +1,5 @@
+#pragma once
+
+
+void setMeanSolarDistance(float newMeanSolarDistance);
+float getMeanSolarDistance();
